# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Oraclous Platform API
  version: 1.0.0
  description: >-
    The public REST contract served by the application-gateway-service — the platform's single
    external edge (ADR-015, ADR-018). Every request is authenticated once at the edge (a bearer JWT
    issued by the identity service); the gateway forwards a verified identity downstream and streams the
    upstream response back. Every 4xx/5xx the caller sees is the canonical **error envelope**
    (`components/schemas/ErrorEnvelope`) — gateway-own errors and normalised upstream errors alike;
    an upstream body is never relayed, so no stack/SQL/internal-host detail can leak (Interface
    Contracts §3).


    **Stability (ADR-015 §2):** each operation carries `x-stability: stable | provisional`. A `stable`
    operation's shape is locked; breaking changes require a new major version. The `openapi-diff-gate`
    CI job blocks breaking changes to `stable` operations.


    **Evolving contract (R6):** this is the R6-in-progress public surface and grows per release slice
    (integration keys + published agents, chat, webhooks, MCP land in later R6 slices and are added
    here as they ship). The published `openapi/v1.yaml` — not any planning table — is the source of
    truth for what the gateway actually exposes.


    **As-built path note:** the proxied paths are currently heterogeneous (`/v1/...`, `/api/v1/...`,
...
    actually route today.


...


...
servers:
  - url: /
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/capabilities:
    get:
      tags: [capabilities]
      summary: The unified capability registry.
      operationId: listCapabilities
      x-stability: stable
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /api/v1/capabilities/match:
    post:
      tags: [capabilities]
      summary: Match capabilities by name — returns the registry entries whose names match.
      operationId: matchCapabilities
      x-stability: stable
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonObject" }
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/capabilities/{capabilityId}:
    get:
      tags: [capabilities]
      summary: Read one capability descriptor (org-scoped; an unknown / cross-org id is a 404 — mask).
      operationId: getCapability
      x-stability: stable
      parameters:
        - { name: capabilityId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The capability descriptor.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DescriptorResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [capabilities]
      summary: Replace a capability's descriptor (org-scoped).
      operationId: updateCapability
      x-stability: stable
      parameters:
        - { name: capabilityId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonObject" }
      responses:
        "200":
          description: The updated capability descriptor.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DescriptorResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      tags: [capabilities]
      summary: Delete a capability descriptor (org-scoped).
      operationId: deleteCapability
      x-stability: stable
      parameters:
        - { name: capabilityId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/integration-keys:
    post:
      tags: [keys]
      summary: Mint an integration key bound to a slug XOR a capability list; the secret is shown once.
      operationId: mintIntegrationKey
      x-stability: provisional
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonObject" }
      responses:
        "201": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    get:
      tags: [keys]
      summary: List the organisation's integration keys (redacted — never the hash or secret).
      operationId: listIntegrationKeys
      x-stability: provisional
      parameters:
        - { $ref: "#/components/parameters/LimitParam" }
        - { $ref: "#/components/parameters/OffsetParam" }
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Unauthorized" }
  /v1/integration-keys/{keyId}:
    get:
      tags: [keys]
      summary: Read one integration key (redacted), org-scoped.
      operationId: getIntegrationKey
      x-stability: provisional
      parameters:
        - { name: keyId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [keys]
      summary: Revoke an integration key (soft tombstone — terminal), org-scoped.
      operationId: revokeIntegrationKey
      x-stability: provisional
      parameters:
        - { name: keyId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Revoked. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/integration-keys/{keyId}/rotate:
    post:
      tags: [keys]
      summary: Rotate a LIVE key's secret in place (the old dies; a revoked key 404s — revoke is final).
      operationId: rotateIntegrationKey
      x-stability: provisional
      parameters:
        - { name: keyId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # --- Member console chat (R6 Slice 6) — gateway-owned, org + creating-member scoped ---
  /v1/chat/threads:
    post:
      tags: [chat]
      summary: Start a chat thread bound to a published agent (member, private to the creator).
      operationId: startChatThread
      x-stability: provisional
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonObject" }
      responses:
        "201": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    get:
      tags: [chat]
      summary: List the member's own chat threads (most-recent first).
      operationId: listChatThreads
      x-stability: provisional
      parameters:
        - { $ref: "#/components/parameters/LimitParam" }
        - { $ref: "#/components/parameters/OffsetParam" }
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Unauthorized" }
  /v1/chat/threads/{threadId}:
    delete:
      tags: [chat]
      summary: Soft-delete a chat thread (member, the creator only).
      operationId: deleteChatThread
      x-stability: provisional
      parameters:
        - { name: threadId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/chat/threads/{threadId}/messages:
    post:
      tags: [chat]
      summary: Send a message — runs the bound agent on the harness + persists the turn.
      operationId: sendChatMessage
      x-stability: provisional
      parameters:
        - { name: threadId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonObject" }
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "5XX": { $ref: "#/components/responses/Upstream5xx" }
    get:
      tags: [chat]
      summary: Read a chat thread's transcript (member, the creator only).
      operationId: listChatMessages
      x-stability: provisional
      parameters:
        - { name: threadId, in: path, required: true, schema: { type: string } }
        - { $ref: "#/components/parameters/LimitParam" }
        - { $ref: "#/components/parameters/OffsetParam" }
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/chat/threads/{threadId}/messages/{messageId}/feedback:
    post:
      tags: [chat]
      summary: Rate an assistant message thumbs up/down (member, the creator only; idempotent).
      operationId: setChatMessageFeedback
      x-stability: provisional
      parameters:
        - { name: threadId, in: path, required: true, schema: { type: string } }
        - { name: messageId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonObject" }
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }

  # --- Inbound webhook ingress (R6 Slice 7) — PUBLIC, signature-gated (no bearer) ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        A short-lived access token minted by the identity service (member credentials). The gateway verifies
        it once at the edge and forwards a verified identity downstream (ADR-018). Integration-key and
        agent-credential auth modes are added in later R6 slices.
  parameters:
    LimitParam:
      name: limit
      in: query
      required: false
      description: >-
        Maximum number of items to return (WP-10 pagination). OPTIONAL and
        backward-compatible: a caller that omits it gets the first 100 items
        (the prior behaviour for any realistic page). Clamped to a hard maximum
        of 200, so no single read is unbounded. The response shape is unchanged
        — a plain array, never an items/total envelope.
      schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
    OffsetParam:
      name: offset
      in: query
      required: false
      description: >-
        Number of items to skip before collecting the page (WP-10 pagination).
        OPTIONAL; defaults to 0 (start of the stable ordering). Paired with a
        deterministic ORDER BY so paging is repeatable.
      schema: { type: integer, minimum: 0, default: 0 }
  schemas:
    JsonObject:
      type: object
      description: A free-form JSON object; the precise request shape is owned by the upstream service.
      additionalProperties: true
...
    DescriptorResponse:
      type: object
      description: >-
        A capability/tool descriptor as returned by the registry. The full body is owned by the
        capability-registry-service; only the load-bearing fields are named here. `status`
        distinguishes an executable descriptor from one awaiting the MCP-import supply-chain gate.
      additionalProperties: true
      properties:
        id: { type: string, format: uuid }
        organisation_id: { type: string, format: uuid }
        kind: { type: string, description: The descriptor kind (e.g. tool). }
        name: { type: [string, "null"] }
        descriptor:
          type: object
          additionalProperties: true
          description: The OHM manifest JSONB.
        status:
          type: string
          enum: [active, pending_approval]
          default: active
          description: >-
            `active` is executable; `pending_approval` is an imported MCP tool awaiting an admin's
            approval (it is not executable until approved).
...
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [field, issue]
      properties:
        field:
          type: string
          minLength: 1
          description: The offending field name or path. Never the value.
        issue:
          type: string
          pattern: "^[A-Z][A-Z0-9_]*$"
          description: Machine-readable issue token (e.g. INVALID_FORMAT). Describes the problem, not the value.
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      description: >-
        Canonical error envelope for every 4xx/5xx the gateway produces (own + normalised upstream).
        `additionalProperties` is false at every level so no stack/cause/exception/trace field can ever
        exist. Mirrors packages/errors/contract/error-envelope.schema.json (the cross-repo source).
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId, retryable]
          properties:
            code:
              type: string
              description: Closed machine-readable taxonomy; the only branch key. Never branch on message.
              enum:
                - VALIDATION_FAILED
                - MALFORMED_REQUEST
                - UNAUTHENTICATED
                - UNAUTHORIZED
                - NOT_FOUND
                - METHOD_NOT_ALLOWED
                - CONFLICT
                - CREDENTIALS_REQUIRED
                - PAYLOAD_TOO_LARGE
                - UNSUPPORTED_MEDIA_TYPE
                - RATE_LIMITED
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
                - GATEWAY_TIMEOUT
            message:
              type: string
              minLength: 1
              description: Curated, generic, human-readable. Never an exception message; never reflects request content.
            requestId:
              type: string
              pattern: "^req_[0-9A-Za-z]+$"
              description: Opaque correlation handle; the only trace exposed. Maps server-side to the full trace.
            retryable:
              type: boolean
              description: Server-authoritative — may a retry succeed? The FE reads this boolean; it never hardcodes which codes are retryable.
            details:
              type: array
              minItems: 1
              description: Present only for VALIDATION_FAILED. Closed sub-vocabulary; never the raw offending value.
              items: { $ref: "#/components/schemas/ErrorDetail" }
            needs_credential:
              type: object
              additionalProperties: false
              required: [requirement_id, provider]
              description: Present only for CREDENTIALS_REQUIRED. Names the missing credential requirement + provider so the FE can prompt onboarding. Leak-safe machine tokens — never a value, URL, host, or secret.
              properties:
                requirement_id:
                  type: string
                  minLength: 1
                  maxLength: 64
                  pattern: "^[A-Za-z0-9][A-Za-z0-9_.-]*$"
                  description: The missing requirement's type token, e.g. api_key. Never a value.
                provider:
                  type: string
                  minLength: 1
                  maxLength: 48
                  pattern: "^[A-Za-z0-9][A-Za-z0-9_.-]*$"
                  description: The provider the credential is for, e.g. web_search. Never a value.
  responses:
    UpstreamJson:
      description: Success — a JSON body produced and owned by the upstream service, streamed through unchanged.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/JsonObject" }
    Unauthenticated:
      description: No or invalid credentials (UNAUTHENTICATED).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Authenticated but not permitted (UNAUTHORIZED).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: No such resource, or an unrouted path (NOT_FOUND).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    ValidationFailed:
      description: The request failed validation (VALIDATION_FAILED; carries details[]).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    Upstream5xx:
      description: An upstream failure normalised to the envelope (SERVICE_UNAVAILABLE / GATEWAY_TIMEOUT / INTERNAL_ERROR).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...